Housekeeping: Renamed repository layer files for clarity
